6. Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -6,9 +6,9 @@
 int main(void)
 {
 print("Size of a char:%c byte(s)\n", sizeof(char));
-printf("Size of an int:%d byte(s)\n", sizeof(int));
-printf("Size of a long int:%ld byte(s)\n", sizeof(long int));
-printf("Size of a long long int:%lld byte(s)\n", sizeof(long long int));
+printf("Size of an int:%d byte(s)\n", (unsigned int)sizeof(int));
+printf("Size of a long int:%ld byte(s)\n", (unsigned int)sizeof(long int));
+printf("Size of a long long int:%lld byte(s)\n", (unsigned int)sizeof(long long int));
 printf("Size of a float:%d byte(s)\n", sizeof(float));
 return (0);
 }
